feat: Add date/time module
On page load, sets the starting time for the event to
the nearest half hour in the future and the end time for
the event to 30 minutes after that. Verifies that end time
is not before start time; if it is, set the end time to 30 minutes
after the start time.