chore: Update gulp build process
Gulp build process now:
* Properly empties dist directory beforehand
* Concatenates js/lib files into app.js
* Uses htmlmin as minify-html was deprecated

Removed a bunch of unneeded tasks as well.

Updated .html files to use mustache.js in app.js.

Added the fetch polyfill well we were at it.

We still use minify-inline instead of htmlmin's options to
minify js because htmlmin doesn't give a nice option for
ignoring script tags with a certain type. We need that for our
inline mustache templates.

In the future, we could work around by pre-compiling mustache
templates in the build process or figuring out the crazy regex
way htmlmin handles this. But for now, this is fine.