feat: Add logged in and log out modules and styles
Logged in module checks the faux backend if there is a logged in
user. It applies a style to root if there is. Added styles to css
to take advantage and hide menu options based on logged in status.

Log out module logs out the user when you click the specified item
then updates the styles.